Relocated logo and added Footer52 component to App
The logo from ProjectScreen.vue was removed and a new Footer52.vue component was created containing the logo and other information. This Footer52 component was added to App.vue serving as the universal footer across the application. Through this, the overall structure and visual consistency of the application has improved.