Fix sorting of patient submissions in PatientScreen
Resolved a bug in the PatientScreen component where submissions were not correctly sorted by their creation date. The sorting operation has been adjusted to ensure that it is performed after the submissions are fetched for a patient. This change will ensure that the submissions are always displayed in the correct chronological order.